refactor(voice): extract shared TTS types to tts_types.py (issue #94 phase 3)
把被 supervisor/synthesis/playback 三方共用的类型与无状态 helper 抽到新模块
app/voice/tts_types.py，为后续拆分避免循环依赖：
- TTSPreparedAudio / _TTSRequest
- TTSServiceState + _set_service_state
- _provider_is_closed / _parse_service_endpoint / TTSCallback

tts.py 通过 import 原样 re-export 这些符号，既有导入路径与全部测试不变。
